refactor(demo): fix base href, move universal scripts to subdir
