+++ demo/src/index.html
@@ -26,7 +26,6 @@
   <meta name="msapplication-navbutton-color" content="#1d1d2b">
   <meta name="apple-mobile-web-app-status-bar-style" content="black-translucent">
   <meta name="msapplication-starturl" content="/">
-  <base href="/">
 </head>
 <body>
   <bs-demo>
