+++ package.json
@@ -41,8 +41,8 @@
     "view-stats": "webpack-bundle-analyzer demo/dist/stats.json",
     "build:dynamic": "npm run build:client-and-server-bundles && npm run webpack:server",
     "serve:dynamic": "node demo/dist/server.js",
-    "build:client-and-server-bundles": "ng build --prod --env=server && ng build --prod --env=server --app 1 --output-hashing=false",
-    "webpack:server": "webpack --config webpack.server.config.js --progress --colors"
+    "build:client-and-server-bundles": "ng build -bh / --prod --env=server && ng build -bh / --prod --env=server --app 1 --output-hashing=false",
+    "webpack:server": "webpack --config ./scripts/universal/webpack.server.config.js --progress --colors"
   },
   "main": "bundles/ngx-bootstrap.umd.js",
   "module": "index.js",
