+++ scripts/universal/static.paths.ts
